backend/requirements.txt
fastapi==0.116.0
uvicorn[standard]==0.34.0
pydantic==2.11.10
SQLAlchemy==2.0.38
alembic==1.13.3
azure-storage-blob==12.25.0
azure-servicebus==7.12.1
pyodbc==5.2.0
anyio==4.9.0
azure-identity==1.19.0
fpdf2==2.8.1
pytest==8.3.1
# Agent framework
semantic-kernel[azure]==1.41.1
# Azure AI services for media extraction
azure-cognitiveservices-speech==1.43.0
azure-ai-vision-imageanalysis==1.0.0
opencv-python-headless==4.11.0.86
# Token estimation and cost tracking
tiktoken==0.9.0
# DOCX export
python-docx==1.1.2
